app/requirements.txt
amqp==2.6.1
billiard==3.6.3.0
celery==4.4.7
Django==2.2.17
django-admin-sortable==2.2.3
django-configurations==2.2
django-extensions==3.1.0
djangorestframework==3.12.2
gunicorn==20.0.4
kombu==4.6.11
psycopg2-binary==2.8.6
pytz==2020.4
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
vine==1.3.0
